third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/B.1.2-expected.txt

PASS Function("'use strict'; \"\\010\"") threw exception of type SyntaxError.
PASS !!Function("\"\\010\"") is true
PASS true === true
PASS Function("'use strict'; \"\\00\"") threw exception of type SyntaxError.
PASS !!Function("\"\\00\"") is true
PASS true === true
PASS Function("'use strict'; \"\\1\"") threw exception of type SyntaxError.
PASS !!Function("\"\\1\"") is true
PASS true === true
PASS Function("'use strict'; \"\\08\"") threw exception of type SyntaxError.
PASS !!Function("\"\\08\"") is true
PASS true === true
PASS !!Function("'use strict'; \"\\0\"") is true
PASS !!Function("\"\\0\"") is true
PASS true === true
PASS !!Function("'use strict'; \"\\0x\"") is true
PASS !!Function("\"\\0x\"") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
